# docs/swagger.yaml
basePath: /api
definitions:
  handlers.ClassificationStatsResponse:
    properties:
      average_confidence:
        type: number
      classified_items:
        type: integer
      pendingItems:
        type: integer
      total_items:
        type: integer
    type: object
  handlers.ClassifyRequest:
    properties:
      category:
        type: string
      classifier:
        type: string
      context:
        additionalProperties: true
        type: object
      item_code:
        type: string
      item_id:
        type: integer
      item_name:
        type: string
      model:
        type: string
      options:
        additionalProperties: true
        type: object
      strategy_id:
        type: string
    type: object
  handlers.ClassifyResponse:
    properties:
      attributes:
        additionalProperties: true
        type: object
      category:
        type: string
      classifier:
        type: string
      code:
        type: string
      confidence:
        type: number
      item_code:
        type: string
      item_id:
        type: integer
      item_name:
        type: string
      name:
        type: string
      result: {}
      subcategory:
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        type: boolean
      message:
        type: string
    type: object
  handlers.MonitoringDataResponse:
    properties:
      providers:
        items:
          $ref: '#/definitions/handlers.ProviderMetricsResponse'
        type: array
      system:
        $ref: '#/definitions/handlers.SystemStatsResponse'
    type: object
...
  handlers.ProviderMetricsResponse:
    properties:
      active_channels:
        type: integer
      average_latency_ms:
        type: number
      current_requests:
        type: integer
      failed_requests:
        type: integer
      id:
        type: string
      last_request_time:
        type: string
      name:
        type: string
      requests_per_second:
        type: number
      status:
        type: string
      successful_requests:
        type: integer
      total_requests:
        type: integer
    type: object
...
  handlers.SystemStatsResponse:
    properties:
      active_providers:
        type: integer
      system_requests_per_second:
        type: number
      timestamp:
        type: string
      total_failed:
        type: integer
      total_providers:
        type: integer
      total_requests:
        type: integer
      total_successful:
        type: integer
    type: object
...
host: localhost:9999
info:
  contact:
    email: support@example.com
    name: API Support
  description: API для системы нормализации данных из 1С. Мульти-провайдерная нормализация,
    AI-классификация, управление качеством данных.
  license:
    name: Internal Use Only
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: HTTP Server API
  version: "1.0"
paths:
...
  /api/classification/classify:
    post:
      consumes:
      - application/json
      description: Классифицирует элемент номенклатуры по справочнику КПВЭД
      parameters:
      - description: Данные для классификации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ClassifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Результат классификации
          schema:
            $ref: '#/definitions/handlers.ClassifyResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Классифицировать элемент
      tags:
      - classification
  /api/classification/stats:
    get:
      consumes:
      - application/json
      description: Возвращает статистику процесса классификации
      parameters:
      - description: ID базы данных
        in: query
        name: database_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Статистика классификации
          schema:
            $ref: '#/definitions/handlers.ClassificationStatsResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получить статистику классификации
      tags:
      - classification
...
  /api/monitoring/providers:
    get:
      consumes:
      - application/json
      description: Возвращает текущий статус и метрики всех AI-провайдеров
      produces:
      - application/json
      responses:
        "200":
          description: Данные мониторинга провайдеров
          schema:
            $ref: '#/definitions/handlers.MonitoringDataResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получить статус AI-провайдеров
      tags:
      - monitoring
  /api/monitoring/providers/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает детальные метрики конкретного AI-провайдера
      parameters:
      - description: ID провайдера
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Метрики провайдера
          schema:
            $ref: '#/definitions/handlers.ProviderMetricsResponse'
        "404":
          description: Провайдер не найден
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получить метрики провайдера
      tags:
      - monitoring
  /api/monitoring/providers/{id}/start:
    post:
      consumes:
      - application/json
      description: Запускает указанный AI-провайдер
      parameters:
      - description: ID провайдера
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешный запуск
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Провайдер не найден
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Запустить провайдер
      tags:
      - monitoring
  /api/monitoring/providers/{id}/stop:
    post:
      consumes:
      - application/json
      description: Останавливает указанный AI-провайдер
      parameters:
      - description: ID провайдера
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешная остановка
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Провайдер не найден
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Остановить провайдер
      tags:
      - monitoring
...
schemes:
- http
- https
swagger: "2.0"